test: fix Ray workers after local_mode removal
- Replace ray.init(local_mode=True) with standard local cluster init in
  service test modules that use Ray actors.
- tests/service_tests/conftest.py: prepend tests/service_tests to PYTHONPATH
  so workers resolve util.* test packages.
- api/api_tests/conftest.py: prepend api/ so workers resolve api_tests.*.
- Move wrap_callable_as_stage tests out of .../util/pipeline/ (cloudpickle
  mis-import as util.pipeline) to test_wrap_callable_as_stage.py; remove
  stray @ray.method blocks from message broker task source tests.

Made-with: Cursor
